% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.muni.r
\docType{data}
\name{cal.muni}
\alias{cal.muni}
\title{Result by municipality in exam Saber 3, 5 and 9 for 9th grade in mathematics test}
\format{This \code{data.frame} has 1125 rows and 6 columns.
\describe{
\item{dpto}{Identificador del departamento}
\item{namesdpto}{Nombre del departamento}
\item{mpio}{Es el identificador del municipio}
\item{namesmpio}{Nombre del municipio}
\item{codETC}{Identificador de la entidad territorial certificada (ETC)}
\item{prom}{Resultado promedio en matem치ticas para grado 9 de
            cada municipio colombiano participante en la prueba Saber 3, 5 y 9
            en el 2015.}}}
\source{
\href{http://www.icfes.gov.co/item/1861-investigador-el-icfes-actualizo-ftp-para-facilitar-acceso-a-sus-bases-de-datos}{ftp://ftp.icfes.gov.co/}
}
\usage{
cal.muni
}
\description{
\code{data.frame} que contiene los puntajes en matem치ticas del grado 9
agregados por municipio, en el examen Saber 3, 5 y 9 aplicadas en el 2015 en Colombia.
}
\references{
Instituto Colombiano para la Evaluaci칩n de la Educaci칩n ICFES, (2016).
\url{http://www.icfes.gov.co/}
}
\keyword{datasets}

