% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.r
\docType{data}
\name{divipola}
\alias{divipola}
\title{The political and administrative division of Colombia}
\format{The file has the following information:
\describe{
\item{id.espacia}{Identifier of the municipality linked to the identifier of the department.}
\item{dpto}{Identifier of the department}
\item{mpio}{Identifier of the municipality}
\item{codETC}{Identifier of the certified territorial entity ((ETC, for its acronym in Spanish))}
\item{tip.etc}{Type of territorial entity: departmental or municipal}}}
\source{
\url{http://geoportal.dane.gov.co:8084/Divipola/}
}
\usage{
divipola
}
\description{
\code{data.frame} which contains the list of political division
 in Colombia (divipola).
}
\references{
National Statistics Department - Departamento Administrativo Nacional
 de Estadística - DANE, (2016). \url{http://www.dane.gov.co/}
}
\keyword{datasets}

