% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.GREG.r
\name{E.GREG}
\alias{E.GREG}
\title{Generalized regression estimator (GREG)}
\usage{
E.GREG(x, Xk, total, prom, weight, stratum = NULL, ...)
}
\arguments{
\item{x}{\code{data.frame} which contains the information necessary to make the estimations.}

\item{Xk}{Vector which contains the names of the auxiliary variables.}

\item{total}{Total of the auxiliary variable.}

\item{prom}{Name of the variable(s) to be estimated.}

\item{weight}{Expansion factor of each observation.}

\item{stratum}{Indicator of the domains in the sample. By default \code{stratum = NULL}.}

\item{...}{Additional arguments for the \code{\link{calib}} function.}
}
\value{
Retorna un \code{data.frame}
\describe{
\item{R2}{El  \eqn{R^2} for the model  defined}
\item{mse}{The mean squared error obtained in the estimation}
}
}
\description{
The function calculates the mean squared error (MSE) for the set of covariables used in the GREG.
}
\examples{
#Reading base
data('ResultStudent')
require(TeachingSampling)
#Selecting data
sampling <- ResultStudent[["student"]]
#Creating dummy variables
sampling <- data.frame(sampling[,c('weight','prop','ses')],
                      Domains(sampling[["urbanicity"]]),
                      Domains(sampling[["schooltype"]]))
#Total variable dummy
total <- ResultStudent[["total"]]

Xk <- c('Rural', 'Urbana','No.Oficial', 'Oficial', 'ses')

E.GREG(x = sampling,total = total, Xk = Xk,
              prom = 'prop',weight = 'weight',method='linear')
}
\references{
\describe{
\item{-}{Särdnal, C. E., Swensson, B., & Wretman, J. H. (1992).
        Model assisted survey sampling.}
\item{-}{Rubin, D. B. (2004). Multiple imputation for
        nonresponse in surveys (Vol. 81). John Wiley & Sons.}
\item{-}{Von Davier, M., Gonzalez, E., & Mislevy, R. (2009).
        What are plausible values and why are they useful. IERI
        monograph series, 2, 9-36.}
}
}
\seealso{
\code{\link{S.fitFH}}, \code{\link{calib}}
}

