% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc2013.r
\docType{data}
\name{etc2013}
\alias{etc2013}
\title{Results of schools in the exam saber 3, 5 and 9, and the exam saber 11.}
\format{An object of class \code{data.frame} with 712 rows and 12 columns.
\describe{
\item{prom}{Average reported to the school by the Colombian Institute for Education Assessment
 (ICFES, for its acronym in Spanish)}
\item{hat.prom}{Estimation of the school’s average}
\item{hat.sd}{Estimation of the standard deviation for the estimated average \code{hat.prom}}
\item{x1}{Result obtained by the school in the 3rd grade Mathematics
          test of the Saber 3, 5, and 9 exams of 2012. }
\item{x2}{Result obtained by the school in the 5th grade Mathematics
          test of the Saber 3, 5, and 9 exams of 2012}
\item{x3}{Result obtained by the school in the 9th grade Mathematics
          test of the Saber 3, 5, and 9 exams of 2012.}
\item{x4}{Result obtained by the school in the 9th grade Mathematics
          test of the Saber 3, 5, and 9 exams of 2013}
\item{x5}{Result obtained by the school in the 11th grade Mathematics
          test of the Saber 11 exam of 2013}
\item{x6}{Result obtained by the school in the11th grade language
          test of Saber 11 exam of 2013}
\item{x7}{Result obtained by the school in the 3rd grade language
          test of the Saber 3, 5, and 9 exams of 2013}
\item{x8}{Result obtained by the school in the 11th grade biology
          test of the Saber 11 exam of 2013 }
\item{x9}{Result obtained by the school in the 11th grade physics
          test of the Saber 11 exam of 2013}}}
\source{
\href{http://www.icfes.gov.co/item/1861-investigador-el-icfes-actualizo-ftp-para-facilitar-acceso-a-sus-bases-de-datos}{ftp://ftp.icfes.gov.co/}
}
\usage{
etc2013
}
\description{
The base relies on the results obtained by different schools linked to
a territorial entity of Colombia.  There are 741 schools registered;
77 out of those are part of the controlled samples. Twelve variables
are documented in each one.
}
\references{
Instituto Colombiano para la Evaluación de la Educación ICFES, (2016).
\url{http://www.icfes.gov.co/}
}
\keyword{datasets}

