% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFH.r
\name{fitFH}
\alias{fitFH}
\title{Small Area Estimation under a Fay Herriot model.}
\usage{
fitFH(x, yhat, Sd.yhat, xk, method = "REML", x.predic = NULL,
  formula = NULL)
}
\arguments{
\item{x}{\code{data.frame} which contains the information necessary to make the estimations.}

\item{yhat}{Name of the column containing the variable’s estimated values.}

\item{Sd.yhat}{Name of the column of estimated standard deviation for  \code{yhat}}

\item{xk}{vector of characters which contains the names of the covariantes.}

\item{method}{Estimation method used by the function \code{\link{eblupFH}}}

\item{x.predic}{Observations on which the forecast will be made}
}
\description{
The function makes the estimation of the best lineal unbiased predictor (EBLUP) based on the Fay Herriot model.
}
\examples{
# Reading data
data('etc2013')
# Add dummys
etc2013[["IND"]] <- ifelse(is.na(etc2013[["hat.prom"]]), 'NO','SI')
# Identify observations with missing data
etc2013[["IMP"]] <- apply(etc2013[,c('x1','x2','x3','x4')], 1,
                         function(x){ifelse(anyNA(x),'IMP','NO_IMP')})
table(etc2013[,c("IND","IMP")])
# Separate observations
sampling <- subset( etc2013,IND == 'SI' & IMP == 'NO_IMP')
# Variable prediction
predic   <- subset(etc2013,!c(IND == 'SI' & IMP == 'NO_IMP'))
# Impute observations prediction
imp <- mice::mice(predic[,c('x1','x2','x3','x4')], method = 'norm.boot')
predic[,c('x1','x2','x3','x4')] <- complete(imp)

FH <- fitFH(x = sampling, yhat  = "hat.prom", Sd.yhat = "hat.sd",
                 xk = c('x1','x2','x3','x4'), x.predic = predic)


}
\references{
\describe{
\item{-}{Fay, R.E. and Herriot, R.A. (1979). Estimation of income
from small places: An application of James-Stein procedures to
census data. Journal of the American Statistical Association 74, 269-277.}
\item{-}{Marhuenda, Y., Morales, D. and Pardo, M.C. (2014). Information
criteria for Fay-Herriot model selection. Computational Statistics and
 Data Analysis 70, 268-280}
\item{-}{Rao, J.N.K. (2003). Small Area Estimation. Wiley, London.}}
}
\seealso{
\code{\link{S.fitFH}, \link{eblupFH}}
}

