% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknifeVar.r
\name{jackknifeVar}
\alias{jackknifeVar}
\title{Estimated variance for the estimator GREG using the jackknife method}
\usage{
jackknifeVar(x, yhat, weight, total, Xk, stratum = NULL, ...)
}
\arguments{
\item{x}{\code{data.frame} que contiene la información necesaria para realizar las estimaciones}

\item{yhat}{Nombre de la columna que contiene los valores estimados de la variable}

\item{weight}{Factor de expansión de cada observación}

\item{total}{Total de la variable auxiliar}

\item{Xk}{Vector que contiene los nombres de las covariables.}

\item{stratum}{Indicadora de los dominios en la muestra, por defecto \code{stratum = NULL}}

\item{...}{Argumentos adicionales empleados por la función \link{calib}}
}
\value{
Retorna un \code{\link{data.frame}} con los resultados obtenidos para el estimador de Horvitz-Thompson  y
el GREG, así como sus respectivas varianzas para los grupos y subgrupos indicados.
}
\description{
La función realiza la estimación de las varianzas para el estimador
Horvitz-Thompson (TH)  y el estimador de regresión generalizada (GREG)
cuando se emplean los valores plausibles. La estimación de la varianza estimada
para un dominio viene dada por la  ecuación
\deqn{EE(\hat{\eta}) = \sqrt{\frac{1}{5}\sum_{i=1}^{5}Var_m(\hat{\eta}_i)+
         \left(1 + \frac{1}{5}\right)\frac{1}{4}\sum_{i=1}^{5}(\hat{\eta}_i -
         \hat{\eta})^2}}

donde \eqn{Var_m(\hat{\eta}_i)} es la varianza del estimador calculada a partir
de las observaciones y el  diseño muestral.
El segundo término de la ecuación es estimada utilizando el método jackknife.
En el caso del  estimador de regresión generalizada, es necesario ajustar un modelo lineal,
donde el valor plausible \eqn{\hat{\eta}_{ij}} es incluido como la variable respuesta, así:
\deqn{\hat{\eta}_{ij}=\beta X_i+e_{ij}}
con \eqn{X_i} es la información auxiliar, empleando los residuales estimados del modelo
se estima la varianza del GREG para cada dominio, utilizando la expresión
\deqn{\widehat{Var}(\hat{\eta}_{GREG})=\sum\sum_s\frac{\Delta_{jl}}{\pi_{jl}}g_{js}
      \frac{\hat{e}_{js}}{\pi_j}g_{ls}\frac{\hat{e}_{ls}}{\pi_l}}
donde
\describe{
\item{-}{\eqn{g_{js} = 1+\left( \bar{X}-\hat{\bar{X}}\right)^t \left(\sum_m \frac{w_j}{\sigma^2_l}x_jx_j^t\right)^{-1}x_j}}
\item{-}{\eqn{\hat{e}_{ij} = \hat{\eta}_{ij} - \hat{\beta} X_i}}
\item{-}{\eqn{w_j = \frac{1}{\pi_j}}}
}
}
\examples{
data('ResultStudent')
attach(ResultStudent)
sampling <- ResultStudent[["student"]]
#Creando variables dummys
sampling <- data.frame(sampling[,c("schooltype",'weight','prop','ses')],
                      Domains(sampling[["urbanicity"]]),
                      Domains(sampling[["schooltype"]]))

#Identifica covariables
Xk <- c('Rural', 'Urbana','No.Oficial', 'Oficial', 'ses')

jackknifeVar(x=sampling,yhat = "prop",weight = 'weight',
            Xk=c("ses", "Rural","Urbana"),
            total = total, method='linear')

jackknifeVar(sampling,weight = 'weight',yhat = "prop",
         Xk = 'ses',total = total, method='linear')
jackknifeVar(sampling,yhat = "prop",weight = 'weight',
           Xk = 'ses',total = total, method='linear',
           stratum = 'schooltype')
}
\references{
\describe{
\item{-}{Shao, J., & Tu, D. (2012). The jackknife and bootstrap. Springer Science & Business Media.}
\item{-}{Rubin, D. B. (2004). Multiple imputation for nonresponse in surveys (Vol. 81). John Wiley & Sons.}
\item{-}{Von Davier, M., Gonzalez, E., & Mislevy, R. (2009). What are plausible values and why are
       they useful. IERI monograph series, 2, 9-36.}
}
}
\seealso{
\code{\link{calib}}
}

