% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMapColombia.r
\name{PlotMapColombia}
\alias{PlotMapColombia}
\title{Plot map of Colombia}
\usage{
PlotMapColombia(x, x.graph, id.nivagre, nivagre, n = 10, divipola, ShapesCol)
}
\arguments{
\item{x}{\code{data.frame}}

\item{x.graph}{Nombre de la variable contenida en \code{x} a graficar en el
el mapa.}

\item{id.nivagre}{Variable que contiene los códigos de identificacion según
la desagregación deseada}

\item{nivagre}{Nivel de agregación deseado municipal ('mpio'),
departamental ('dpto') o entidad territorial certificada ('ETC')}

\item{n}{Número de categorías en que se divide variable \code{x.graph}}

\item{divipola}{\code{data.frame} que relaciona la división política en Colombia.}

\item{ShapesCol}{Shapes de la división político administrativa de Colombia}
}
\value{
El mapa de Colombia como objeto ggplot
}
\description{
Función que dibuja el mapa de Colombia y los resultados agregados por nivel deseado.
}
\examples{
data('cal.muni')
data(divipola)
data(ShapesCol)
PlotMapColombia(x = cal.muni[,c('mpio','prom')], x.graph = 'prom',
               id.nivagre = 'mpio', nivagre = 'mpio',
               divipola = divipola, ShapesCol = ShapesCol)

dpto.prom <- aggregate(cal.muni[,'prom'], list(dpto = cal.muni[['dpto']]), mean)

PlotMapColombia(x = dpto.prom, x.graph = 'x', id.nivagre = 'dpto',nivagre = 'dpto',
               divipola = divipola, ShapesCol = ShapesCol)

etc.prom <-aggregate(cal.muni[,'prom'],list(etc = cal.muni[['codETC']]), mean)

mapCol <- PlotMapColombia(x = etc.prom,x.graph = 'x',id.nivagre = 'etc',
                         nivagre = 'ETC',divipola = divipola,
                         ShapesCol = ShapesCol)

subSanAndres <- mapCol + xlim(c(165298, 171000)) + ylim(c(1884070, 1897000)) +
               guides(fill = "none") + labs(title = "")

subCatalina <- mapCol + xlim(c(205000, 210389)) + ylim(c(1975000, 1984871)) +
guides(fill = "none") + labs(title = "")

require(grid)
grid.newpage()
print(mapCol)
vp <- viewport(width = 0.1, height = 0.2, x = 0.02, y = 0.75, just = c("left", "bottom"))
print(subSanAndres, vp = vp)

vp <- viewport(width = 0.1, height = 0.2, x = 0.15, y = 0.75, just = c("left", "bottom"))
print(subCatalina, vp = vp)

}
\seealso{
\code{\link{ggmap}, \link{ggplot2}, \link{divipola}, \link{etc2013},\link{ShapesCol}}
}

