% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.fitFH.r
\name{S.fitFH}
\alias{S.fitFH}
\title{Model Selection Fay Herriot based on the CME}
\usage{
S.fitFH(x, yhat, Sd.yhat, Xk, p = 1, y = NULL)
}
\arguments{
\item{x}{\code{data.frame} which contains the model´s variables.}

\item{yhat}{Name of the column in \code{x} which indicates the variable to estimate.}

\item{Sd.yhat}{Name of the column in \code{x} which indicates the standard deviation
estimated for \code{yhat}.}

\item{Xk}{Vector which contains the names of the auxiliary variables. These must
be contained in \code{x}.}

\item{p}{Maximum number of combinations to make with the covariantes.
By default \code{p = 1}.}

\item{y}{Optional. Name of the column in \code{x} that has the actual values
of the estimated variable.}
}
\value{
Returns an object of the \code{data.frame} class, which has the summary measures
employed for the selection of the desired model.
\describe{
\item{model}{List of models resulting from \eqn{\sum_{i=1}^{p} {{p}\choose{i}}}}
\item{loglike}{Quality measures related to the model based on the likelihood function}
\item{mse.Sample}{Mean squared error obtained for observations without missing data in their covariantes.}
\item{N.sample}{Number of observations without missing data.}
\item{mse.predict}{Mean squared error obtained for the forecasts.}
\item{N.predict}{Number of observations with prediction}
\item{mse.predictIMP}{Mean square error obtained for predictions with observations missing covariates}
\item{N.predictIMP}{Number of observations with missing data}
}
}
\description{
Function that selects a Fay Herriot model based on the mean squared error (MSE),
based on all possible combinations of \code{p} variables
}
\examples{
data('etc2013')
Xk <- c('x1','x2','x3','x4','x5','x6','x7','x8','x9')
Sel.fit <- S.fitFH(etc2013, 'hat.prom', 'hat.sd',Xk, p = 2)
Sel.fit <- S.fitFH(etc2013, 'hat.prom', 'hat.sd',Xk, p = 2, y = 'prom')
}
\references{
\describe{
\item{-}{Fay, R.E. and Herriot, R.A. (1979). Estimation of income
from small places: An application of James-Stein procedures to
census data. Journal of the American Statistical Association 74, 269-277.}
\item{-}{Marhuenda, Y., Morales, D. and Pardo, M.C. (2014). Information
criteria for Fay-Herriot model selection. Computational Statistics and
 Data Analysis 70, 268-280}
\item{-}{Rao, J.N.K. (2003). Small Area Estimation. Wiley, London.}}
}

