% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.fitGREG.r
\name{S.fitGREG}
\alias{S.fitGREG}
\title{Model selection  for the GREG based on CME}
\usage{
S.fitGREG(x, Xk, total, prom, weight, p = 1, stratum = NULL, ...)
}
\arguments{
\item{x}{\code{data.frame} which contains the model’s variables.}

\item{Xk}{Vector of characters with the names of the co-variables to
be used in the model. These must be contained in the columns of  \code{x}.}

\item{total}{Total of the auxiliary variable.}

\item{prom}{Name of the column in \code{x} which indicates the variable to estimate.}

\item{weight}{Expansion factor of each observation.}

\item{p}{Maximum number of combinations to make with the covariantes. By default \code{p = 1}.}

\item{stratum}{Indicator of the domains in the sample. By default  \code{stratum = NULL}}

\item{...}{Argumentos adicional empleado por la función \code{\link{calib}}.}
}
\value{
Returns a \code{data.frame} where report is made.
\describe{
\item{model}{List of models resulting from \eqn{\sum_{i=1}^{p} {{p}\choose{i}}}}
\item{R2}{El  \eqn{R^2} for each model definedo.}
\item{mse}{Mean squared error obtained in the estimation}}
}
\description{
Function which selects a model for the GREG based on the mean squared error (MSE),
based on all possible combinations that may be obtained from a set of \code{p} variables.
}
\examples{
data('ResultStudent')
attach(ResultStudent)
sampling <- data.frame(student[,c('weight','prop','ses')],
                      Domains(student[["urbanicity"]]),
                      Domains(student[["schooltype"]]))
sampling[["sel"]] <- gl(11,31)
Xk<-c('Rural', 'Urbana','No.Oficial', 'Oficial', 'ses')

S.fitGREG(x = sampling,total = total,Xk = Xk, prom = 'prop',
         weight = 'weight', method = 'linear', p = 3)
S.fitGREG(x = sampling,total = total,Xk = Xk, prom = "prop",
         weight = 'weight',stratum='sel', method = 'linear',p=3)
}
\seealso{
\code{\link{E.GREG}}
}

