% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapesCol.r
\docType{data}
\name{ShapesCol}
\alias{ShapesCol}
\title{Shapes de Colombia}
\format{An object of class \code{list} of length 2.}
\source{
\url{http://sigotn.igac.gov.co/sigotn/default.aspx}
}
\usage{
ShapesCol
}
\description{
Shapes de la División político administrativa de Colombia, descargado del
 Instituto Geográfico Agutín Codazzi (IGAC)
}
\keyword{datasets}

