% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.r
\docType{data}
\name{divipola}
\alias{divipola}
\title{The political and administrative division of Colombia}
\format{El arcivo  cuenta con la siguiente información
\describe{
\item{id.espacia}{Identificador del municipio concatenado con el
                  identificador del departamento}
\item{dpto}{Identificador del departamento}
\item{mpio}{Identificador del municipio}
\item{codETC}{Identificador de la entidad territorial certificada (ETC)}
\item{tip.etc}{Tipo de entidad territorial: departamental o municipal}}}
\source{
\url{http://geoportal.dane.gov.co:8084/Divipola/}
}
\usage{
divipola
}
\description{
\code{data.frame} que contiene la relación de la división política
en Colombia (divipola).
}
\references{
Departamento Administrativo Nacional de Estadística - DANE, (2016).
 \url{http://www.dane.gov.co/}
}
\keyword{datasets}

