% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.GREG.r
\name{E.GREG}
\alias{E.GREG}
\title{Generalized regression estimator (GREG)}
\usage{
E.GREG(x, Xk, total, prom, weight, stratum = NULL, ...)
}
\arguments{
\item{x}{\code{data.frame} que contiene la información necesaria
para realizar las estimaciones.}

\item{Xk}{Vector que contiene los nombres de las variables auxiliares.}

\item{total}{Total de la variable auxiliar.}

\item{prom}{Nombre de la(s) variables(s) a estimar.}

\item{weight}{Factor de expansión de cada observación.}

\item{stratum}{Indicadora de los dominios en la muestra, por defecto \code{stratum = NULL}.}

\item{...}{Argumentos adicionales para la función \code{\link{calib}}.}
}
\value{
Retorna un \code{data.frame}
\describe{
\item{R2}{El  \eqn{R^2} para el modelo definido.}
\item{cme}{El cuadrado medio del error obtenido en la estimación}
}
}
\description{
La función hace el cálculo del error cuadrático medio (ECM) para el conjunto
de covariables usadas el GREG
}
\examples{
#Lectura de la base
data('ResultStudent')
require(TeachingSampling)
#Selección de los datos
sampling <- ResultStudent[["student"]]
#Creando variables dummys
sampling <- data.frame(sampling[,c('weight','prop','ses')],
                      Domains(sampling[["urbanicity"]]),
                      Domains(sampling[["schooltype"]]))
#Total por variable dummy
total <- ResultStudent[["total"]]

#Identifica covariables
Xk <- c('Rural', 'Urbana','No.Oficial', 'Oficial', 'ses')

E.GREG(x = sampling,total = total, Xk = Xk,
              prom = 'prop',weight = 'weight',method='linear')
}
\references{
\describe{
\item{-}{Särdnal, C. E., Swensson, B., & Wretman, J. H. (1992).
        Model assisted survey sampling.}
\item{-}{Rubin, D. B. (2004). Multiple imputation for
        nonresponse in surveys (Vol. 81). John Wiley & Sons.}
\item{-}{Von Davier, M., Gonzalez, E., & Mislevy, R. (2009).
        What are plausible values and why are they useful. IERI
        monograph series, 2, 9-36.}
}
}
\seealso{
\code{\link{S.fitFH}}, \code{\link{calib}}
}

