% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFH.r
\name{fitFH}
\alias{fitFH}
\title{Small Area Estimation under a Fay Herriot model.}
\usage{
fitFH(x, yhat, Sd.yhat, xk, method = "REML", x.predic = NULL)
}
\arguments{
\item{x}{\code{data.frame} que contiene la información necesaria para realizar las estimaciones}

\item{yhat}{Nombre de la columna que contiene los valores estimados de la variable}

\item{Sd.yhat}{Nombre de la columna de desviación estándar estimada para  \code{yhat}}

\item{xk}{Vector de caractres que contiene los nombres de las covariables}

\item{method}{Método de estimación empledo por la función \code{\link{eblupFH}}}

\item{x.predic}{Observaciones a las cuales se realizará el pronóstico}
}
\description{
La función realiza la estimación del mejor predictor lineal  insesgado (EBLUP)
basado en el modelo Fay Herriot.
}
\examples{
# Lectura de los datos
data('etc2013')
# Agregar dummys
etc2013[["IND"]] <- ifelse(is.na(etc2013[["hat.prom"]]), 'NO','SI')
# Identificar observaciones con datos faltantes
etc2013[["IMP"]] <- apply(etc2013[,c('x1','x2','x3','x4')], 1,
                         function(x){ifelse(anyNA(x),'IMP','NO_IMP')})
table(etc2013[,c("IND","IMP")])
# Separar las observaciones
sampling <- subset( etc2013,IND == 'SI' & IMP == 'NO_IMP')
# Variable de pronóstico
predic   <- subset(etc2013,!c(IND == 'SI' & IMP == 'NO_IMP'))
# Imputar observaciones en para el prónostico
imp <- mice::mice(predic[,c('x1','x2','x3','x4')], method = 'norm.boot')
predic[,c('x1','x2','x3','x4')] <- complete(imp)

FH <- fitFH(x = sampling, yhat  = "hat.prom", Sd.yhat = "hat.sd",
                 xk = c('x1','x2','x3','x4'), x.predic = predic)


}
\references{
\describe{
\item{-}{Fay, R.E. and Herriot, R.A. (1979). Estimation of income
from small places: An application of James-Stein procedures to
census data. Journal of the American Statistical Association 74, 269-277.}
\item{-}{Marhuenda, Y., Morales, D. and Pardo, M.C. (2014). Information
criteria for Fay-Herriot model selection. Computational Statistics and
 Data Analysis 70, 268-280}
\item{-}{Rao, J.N.K. (2003). Small Area Estimation. Wiley, London.}}
}
\seealso{
\code{\link{S.fitFH}, \link{eblupFH}}
}

