% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknifeVar.r
\name{jackknifeVar}
\alias{jackknifeVar}
\title{Estimated variance for the estimator GREG using the jackknife method}
\usage{
jackknifeVar(x, yhat, weight, total, Xk, stratum = NULL, ...)
}
\arguments{
\item{x}{\code{data.frame} which contains the information necessary to make the estimations}

\item{yhat}{Name of the column which contains the estimated values of the variable}

\item{weight}{Expansion factor of each observation}

\item{total}{Total for the auxiliary variable}

\item{Xk}{Vector which contains the names of the covariantes.}

\item{stratum}{Indicator of the domains in the sample, by default  \code{stratum = NULL}}

\item{...}{Additional arguments employed by the \link{calib} function.}
}
\value{
Returns a \code{\link{data.frame}} with the results obtained for the Horvitz-Thompson
estimator and the GREG as well as their corresponding variances for the indicated groups and subgroups
}
\description{
The function makes the estimation of variances for the Horvitz-Thompson estimator (TH)
and the generalized regression estimator (GREG) when plausible values are employed.
The estimation of the variance for a domain is given by the equation
\deqn{EE(\hat{\eta}) = \sqrt{\frac{1}{5}\sum_{i=1}^{5}Var_m(\hat{\eta}_i)+
         \left(1 + \frac{1}{5}\right)\frac{1}{4}\sum_{i=1}^{5}(\hat{\eta}_i -
         \hat{\eta})^2}}

where \eqn{Var_m(\hat{\eta}_i)} is the variance of the estimator calculated
from the observations and the sampling design. The second term of the equation
is estimated by using the jackknife method. In the case of the generalized
regression estimator, it is necessary to adjust a lineal model, where
the plausible value \eqn{\hat{\eta}_{ij}} included as the response variable, as follows:
\deqn{\hat{\eta}_{ij}=\beta X_i+e_{ij}}
The GREG variance is estimated for each doman with \eqn{X_i} as auxiliary information,
employing the model’s estimated residuals using the expression

\deqn{\widehat{Var}(\hat{\eta}_{GREG})=\sum\sum_s\frac{\Delta_{jl}}{\pi_{jl}}g_{js}
      \frac{\hat{e}_{js}}{\pi_j}g_{ls}\frac{\hat{e}_{ls}}{\pi_l}}
donde
\describe{
\item{-}{\eqn{g_{js} = 1+\left( \bar{X}-\hat{\bar{X}}\right)^t \left(\sum_m \frac{w_j}{\sigma^2_l}x_jx_j^t\right)^{-1}x_j}}
\item{-}{\eqn{\hat{e}_{ij} = \hat{\eta}_{ij} - \hat{\beta} X_i}}
\item{-}{\eqn{w_j = \frac{1}{\pi_j}}}
}
}
\examples{
data('ResultStudent')
attach(ResultStudent)
sampling <- ResultStudent[["student"]]
#Add dummys
sampling <- data.frame(sampling[,c("schooltype",'weight','prop','ses')],
                      Domains(sampling[["urbanicity"]]),
                      Domains(sampling[["schooltype"]]))

# Covariantes
Xk <- c('Rural', 'Urbana','No.Oficial', 'Oficial', 'ses')

jackknifeVar(x=sampling,yhat = "prop",weight = 'weight',
            Xk=c("ses", "Rural","Urbana"),
            total = total, method='linear')

jackknifeVar(sampling,weight = 'weight',yhat = "prop",
         Xk = 'ses',total = total, method='linear')
jackknifeVar(sampling,yhat = "prop",weight = 'weight',
           Xk = 'ses',total = total, method='linear',
           stratum = 'schooltype')
}
\references{
\describe{
\item{-}{Shao, J., & Tu, D. (2012). The jackknife and bootstrap. Springer Science & Business Media.}
\item{-}{Rubin, D. B. (2004). Multiple imputation for nonresponse in surveys (Vol. 81). John Wiley & Sons.}
\item{-}{Von Davier, M., Gonzalez, E., & Mislevy, R. (2009). What are plausible values and why are
       they useful. IERI monograph series, 2, 9-36.}
}
}
\seealso{
\code{\link{calib}, \link{E.GREG}}
}

