% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.fitFH.r
\name{S.fitFH}
\alias{S.fitFH}
\title{Model Selection Fay Herriot based on the CME}
\usage{
S.fitFH(x, yhat, Sd.yhat, Xk, p = 1, y = NULL)
}
\arguments{
\item{x}{\code{data.frame} que contiene las variables del modelo.}

\item{yhat}{Nombre de la columna en \code{x} que indica la variable ha estimar.}

\item{Sd.yhat}{Nombre de la columna en \code{x} que indica la desviación estándar
estimada para \code{yhat}.}

\item{Xk}{Vector que contiene los nombres de las variables auxiliares,
estas deben estar contenidas en \code{x}.}

\item{p}{Número máximo de combinaciones a realizar con las covariables,
por defecto \code{p = 1}.}

\item{y}{Opcional, nombre de la columna en \code{x} que tiene los valores
reales de la variable estimada.}
}
\value{
Regresa un objeto de clase \code{data.frame}, que tiene las medidas de resumen
empleadas para la selección del modelo deseado.
\describe{
\item{model}{Lista de modelos resultantes de la combinatoria \eqn{\sum_{i=1}^{p} {{p}\choose{i}} } para \eqn{i=1\ldots p}}
\item{loglike}{Medidas de calidad relativas al modelo basado en la función de verosimilitud}
\item{cme.Sample}{Cuadrado medio del error obtenido para las observaciones
                 sin datos faltantes en sus covariables}
\item{N.sample}{Número de observaciones sin datos faltantes}
\item{cme.predict}{Cuadrado medio del error obtenido para los pronósticos}
\item{N.predict}{Número de observaciones a la que se les hace el prónostico}
\item{cme.predictIMP}{El cuadrado medio del error obtenido para los prónosticos
                     con imputación}
\item{N.predictIMP}{Número de observaciones imputadas}}
}
\description{
Función que seleciona un modelo Fay Herriot  basado
en el cuadrado medio del error (CME), con base en todas las posibles
combinanciones de \code{p} variables.
}
\examples{
data('etc2013')
Xk <- c('x1','x2','x3','x4','x5','x6','x7','x8','x9')
Sel.fit <- S.fitFH(etc2013, 'hat.prom', 'hat.sd',Xk, p = 2)
Sel.fit <- S.fitFH(etc2013, 'hat.prom', 'hat.sd',Xk, p = 2, y = 'prom')
}
\references{
\describe{
\item{-}{Fay, R.E. and Herriot, R.A. (1979). Estimation of income
from small places: An application of James-Stein procedures to
census data. Journal of the American Statistical Association 74, 269-277.}
\item{-}{Marhuenda, Y., Morales, D. and Pardo, M.C. (2014). Information
criteria for Fay-Herriot model selection. Computational Statistics and
 Data Analysis 70, 268-280}
\item{-}{Rao, J.N.K. (2003). Small Area Estimation. Wiley, London.}}
}

