% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.fitGREG.r
\name{S.fitGREG}
\alias{S.fitGREG}
\title{Model selection  for the GREG based on CME}
\usage{
S.fitGREG(x, Xk, total, prom, weight, p = 1, stratum = NULL, ...)
}
\arguments{
\item{x}{\code{data.frame} que contiene las variables del modelo}

\item{Xk}{Vector de carácteres con los nombres de las covariables a utilizar en el
modelo, estas deben estar contenidas en las columnas de \code{x}.}

\item{total}{Total de la variable auxiliar}

\item{prom}{Nombre de la columna en \code{x} que indica la variable ha estimar.}

\item{weight}{Factor de expansión de cada observación}

\item{p}{Númere máximo de combinaciones a realizar con las covariables,
por defecto \code{p = 1}.}

\item{stratum}{Indicadora de los dominios en la muestra, por defecto \code{stratum = NULL}}

\item{...}{Argumentos adicional empleado por la función \code{\link{calib}}.}
}
\value{
Regresa un \code{data.frame} donde se reporta
\describe{
\item{model}{Lista de modelos resultantes de la combinatoria de 1 a p}
\item{R2}{El  \eqn{R^2} para cada modelo definido.}
\item{cme}{El cuadrado medio del error obtenido en la estimación}}
}
\description{
Función que seleciona un modelo para el GREG  basado
en el cuadrado medio del error (CME), con base en todas las posibles
combinanciones que puedan ser obtenidas de un  conjunto de \code{p} variables.
}
\examples{
data('ResultStudent')
attach(ResultStudent)
sampling <- data.frame(student[,c('weight','prop','ses')],
                      Domains(student[["urbanicity"]]),
                      Domains(student[["schooltype"]]))
sampling[["sel"]] <- gl(11,31)
Xk<-c('Rural', 'Urbana','No.Oficial', 'Oficial', 'ses')

S.fitGREG(x = sampling,total = total,Xk = Xk, prom = 'prop',
         weight = 'weight', method = 'linear', p = 3)
S.fitGREG(x = sampling,total = total,Xk = Xk, prom = "prop",
         weight = 'weight',stratum='sel', method = 'linear',p=3)
}
\seealso{
\code{\link{E.GREG}}
}

