% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapesCol.r
\docType{data}
\name{ShapesCol}
\alias{ShapesCol}
\title{Shapes de Colombia}
\format{An object of class \code{list} of length 2.}
\source{
\url{http://sigotn.igac.gov.co/sigotn/default.aspx}
}
\usage{
ShapesCol
}
\description{
Shapes of Colombia’s political-administrative division, downloaded from Instituto Geografico Agustin
Codazzi (IGAC, Spanish acronym for Agustin Codazzi Geographic Institute)
}
\keyword{datasets}

